Split coordinator class
